Everything works as it should

+++ Cursus/Libft/ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: akabbadj <akabbadj@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/14 03:42:23 by akabbadj          #+#    #+#             */
-/*   Updated: 2022/11/02 23:03:59 by akabbadj         ###   ########.fr       */
+/*   Updated: 2022/11/04 01:28:21 by akabbadj         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
@@ -14,7 +14,7 @@
 char	*ft_strdup(const char *s1)
 {
 	char	*res;
-	int		slen;
+	size_t	slen;
 
 	slen = ft_strlen(s1) + 1;
 	res = malloc(slen * sizeof(char));
